Remove GoogleTest related code and feature files
The GoogleTest related files and feature were removed because they are no longer relevant to the project. Changes were also made to the ApiCrudSteps file to update the user registration details and confirm successful registration, login, update and deletion messages for the user. The CukesRunner file also updated the dry run value to false, enabling the tests to run. The test results in cucumber.json file were updated as well for completed tests, this was reflected in the cucumber-reports.html update. Lastly, the target rerun text file was updated to not rerun the completed tests.